requirements_emotion.txt
opencv-python
tensorflow
keras
numpy
pillow
